Modify build system and file structure
Replaces the custom build process (which used make) with Hugo Pipes and
Yarn. Hugo Pipes didn't exist when the original build process was put in
place, but it's much simpler and eliminates the need for a lot of other
complexity.

Also, modifies the file structure so that it matches the default for
most Hugo projects. I had originally modified away from Hugo's default
file structure because it was unwieldy when using make, but that's no
longer an issue and it's likely better to be using the default
structure.