Add a proper font stack
The serif and sans-serif stacks in this commit exclusively use various
similar system fonts, so that most text on the site will render without
the need to download anything.

The logo does depend on a web font - Merriweather - but it downloads a
custom version that I have subset exclusively for the letters "N", "W",
and "J" and then compressed using woff2. The result is very small
download, that changes the appearance of the logo and nothing else.

This approach prioritizes minimal site size and thus page load
performance, while still achieving the desired style.